Move the install destination of kinematics utility header
Makes it better aligned with the chosen namespace
