+++ utils/CMakeLists.txt
@@ -3,12 +3,12 @@ add_library(EDM4HEP::kinematics ALIAS kinematics)
 
 target_include_directories(kinematics
   INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-            $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}/utils>)
+            $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>)
 
 target_link_libraries(kinematics PUBLIC INTERFACE ROOT::Core)
 target_compile_features(kinematics INTERFACE cxx_std_17)
 
-install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/ DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/utils
+install(DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/include/ DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
   PATTERN "CMakeLists.txt" EXCLUDE)
